modified Post controller

+++ src/controllers/post.js
@@ -81,12 +81,14 @@ export default class PostController {
    * @returns {object} Success message
    */
   static async deletePost(req, res) {
-    const { id } = req.params;
     try {
+      const { id } = req.params;
+      const userId = req.decoded.user.id;
       const { error } = validateId({ id });
       if (error) return res.status(400).json({ status: 400, error: error.message });
       const Post = await getPost(id);
       if (!Post) return res.status(404).json({ status: 404, error: "Post not found." });
+      if (Post.userId !== userId) return res.status(401).json({ status: 401, error: "Access Denied" });
       await deletePost(id);
       return res.status(200).json({
         status: 200,
@@ -134,10 +136,12 @@ export default class PostController {
   static async updatePost(req, res) {
     try {
       const { id } = req.params;
+      const userId = req.decoded.user.id;
       const { error } = validateId({ id });
       if (error) return res.status(400).json({ status: 400, error: error.message });
       const Post = await PostServices.getPost(id);
       if (!Post) return res.status(404).json({ status: 404, error: "Post not found." });
+      if (Post.userId !== userId) return res.status(401).json({ status: 401, error: "Access Denied" });
       const newPost = await updatePost(id, req.body);
       return res.status(200).json({
         status: 200,
